Inline test packages.
