+++ rdbms/src/test/java/io/spine/server/storage/jdbc/DataSourceWrapperShould.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.storage.jdbc;
+
+import io.spine.server.storage.jdbc.DataSourceWrapper;
+import io.spine.server.storage.jdbc.GivenDataSource;
+import org.junit.Test;
+import io.spine.server.storage.jdbc.DatabaseException;
+
+import static org.junit.Assert.fail;
+import static org.mockito.Mockito.doThrow;
+
+/**
+ * @author Andrey Lavrov
+ */
+public class DataSourceWrapperShould {
+
+    @Test
+    public void throw_database_exception_if_fail_to_close() throws Exception {
+        final GivenDataSource.ClosableDataSource dataSource = GivenDataSource.whichIsAutoCloseable();
+        doThrow(new Exception("")).when(dataSource).close();
+        final DataSourceWrapper wrapper = DataSourceWrapper.wrap(dataSource);
+
+        try {
+            wrapper.close();
+        } catch (DatabaseException expected) {
+            return;
+        }
+        fail();
+    }
+}
