+++ rdbms/src/test/java/io/spine/server/storage/jdbc/DefaultDataSourceConfigConverterShould.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.storage.jdbc;
+
+import com.google.common.testing.NullPointerTester;
+import io.spine.server.storage.jdbc.DataSourceConfig;
+import io.spine.server.storage.jdbc.DefaultDataSourceConfigConverter;
+import org.junit.Test;
+
+import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static org.mockito.Mockito.mock;
+
+/**
+ * @author Dmytro Dashenkov
+ */
+public class DefaultDataSourceConfigConverterShould {
+
+    @Test
+    public void have_private_utility_constructor() {
+        assertHasPrivateParameterlessCtor(DefaultDataSourceConfigConverter.class);
+    }
+
+    @Test
+    public void pass_null_tolerance_test() {
+        final NullPointerTester tester = new NullPointerTester();
+        tester.setDefault(DataSourceConfig.class, mock(DataSourceConfig.class));
+        tester.testStaticMethods(DefaultDataSourceConfigConverter.class,
+                                 NullPointerTester.Visibility.PACKAGE);
+    }
+}
