--- rdbms/src/test/java/io/spine/server/storage/jdbc/aggregate/query/SelectByIdSortedByTimeDescQueryShould.java
@@ -1,62 +1,0 @@
-/*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.storage.jdbc.aggregate.query;
-
-import io.spine.server.storage.jdbc.query.SelectByIdSortedByTimeDescQuery;
-import org.junit.Test;
-import org.slf4j.Logger;
-import io.spine.server.storage.jdbc.DatabaseException;
-import io.spine.server.storage.jdbc.DataSourceWrapper;
-
-import java.sql.SQLException;
-
-import static org.junit.Assert.fail;
-import static org.mockito.ArgumentMatchers.anyBoolean;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-/**
- * @author Dmytro Dashenkov
- */
-public class SelectByIdSortedByTimeDescQueryShould {
-
-    @Test
-    public void throw_DatabaseException_upon_SQL_failure() {
-        final DataSourceWrapper dataSourceMock = mock(DataSourceWrapper.class);
-        when(dataSourceMock.getConnection(anyBoolean())).thenThrow(SQLException.class);
-        final Logger loggerMock = mock(Logger.class);
-
-        final SelectByIdSortedByTimeDescQuery query =
-                SelectByIdSortedByTimeDescQuery.newBuilder("table-name")
-                                               .setDataSource(dataSourceMock)
-                                               .setLogger(loggerMock)
-                                               .build();
-        try {
-            query.execute();
-            fail();
-        } catch (DatabaseException dbe) {
-            verify(loggerMock).error(anyString(), eq(dbe.getCause()));
-        }
-    }
-}
