--- rdbms/src/test/java/io/spine/server/storage/jdbc/entity/query/SelectEntityByIdQueryShould.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.storage.jdbc.entity.query;
-
-import com.google.protobuf.Message;
-import io.spine.server.storage.jdbc.query.SelectEntityByIdQuery;
-import org.junit.Test;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import static org.junit.Assert.assertNull;
-import static org.mockito.Mockito.mock;
-
-/**
- * @author Dmytro Dashenkov
- */
-public class SelectEntityByIdQueryShould {
-
-    @Test
-    public void return_null_if_serialized_message_column_is_empty() throws SQLException {
-        final ResultSet resultSet = mock(ResultSet.class);
-
-        final SelectEntityByIdQuery<String> query =
-                SelectEntityByIdQuery.<String>newBuilder("dummy-table")
-                                     .build();
-
-        final Message deserialized = query.readMessage(resultSet);
-        assertNull(deserialized);
-    }
-}
