--- rdbms/src/test/java/io/spine/server/storage/jdbc/projection/query/SelectTimestampQueryShould.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.storage.jdbc.projection.query;
-
-import io.spine.server.storage.jdbc.query.SelectTimestampQuery;
-import org.junit.Test;
-import io.spine.server.storage.jdbc.DatabaseException;
-
-import java.sql.SQLException;
-
-import static org.junit.Assert.fail;
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.anyString;
-import static org.mockito.Mockito.verify;
-
-/**
- * @author Andrey Lavrov
- */
-public class SelectTimestampQueryShould {
-
-    @Test
-    public void handle_sql_exception() throws SQLException {
-        final SelectTimestampQuery query = Given.getSelectTimestampQueryMock();
-        try {
-            query.execute();
-            fail();
-        } catch (DatabaseException expected) {
-            verify(Given.getLoggerMock()).error(anyString(), any(SQLException.class));
-        }
-    }
-}
