--- rdbms/src/test/java/io/spine/server/storage/jdbc/util/DbTableNameFactoryShould.java
@@ -1,63 +1,0 @@
-/*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.storage.jdbc.util;
-
-import com.google.common.testing.NullPointerTester;
-import com.google.protobuf.StringValue;
-import io.spine.server.entity.AbstractEntity;
-import io.spine.server.storage.jdbc.DbTableNameFactory;
-import org.junit.Test;
-
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
-import static org.junit.Assert.assertEquals;
-
-/**
- * @author Alexander Litus
- */
-@SuppressWarnings("InstanceMethodNamingConvention")
-public class DbTableNameFactoryShould {
-
-    @Test
-    public void have_private_utility_constrctor() {
-        assertHasPrivateParameterlessCtor(DbTableNameFactory.class);
-    }
-
-    @Test
-    public void pass_null_tolerance_check() {
-        final NullPointerTester tester = new NullPointerTester();
-        tester.testStaticMethods(DbTableNameFactory.class, NullPointerTester.Visibility.PACKAGE);
-    }
-
-    @Test
-    public void provide_table_name_for_class() {
-        final String tableName = DbTableNameFactory.newTableName(TestEntity.class);
-
-        assertEquals("google_protobuf_stringvalue",
-                     tableName);
-    }
-
-    private static class TestEntity extends AbstractEntity<String, StringValue> {
-
-        private TestEntity(String id) {
-            super(id);
-        }
-    }
-}
