--- rdbms/src/test/java/io/spine/server/storage/jdbc/util/SerializerShould.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.server.storage.jdbc.util;
-
-import com.google.protobuf.StringValue;
-import io.spine.server.storage.jdbc.Serializer;
-import org.junit.Test;
-
-import static io.spine.Identifier.newUuid;
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
-/**
- * @author Alexander Litus
- */
-@SuppressWarnings("InstanceMethodNamingConvention")
-public class SerializerShould {
-
-    @Test
-    public void have_private_utility_constructor() {
-        assertHasPrivateParameterlessCtor(Serializer.class);
-    }
-
-    @Test
-    public void serialize_and_deserialize_message() {
-        final StringValue expected = StringValue
-                .getDefaultInstance().toBuilder().setValue(newUuid()).build();
-
-        final byte[] bytes = Serializer.serialize(expected);
-        assertTrue(bytes.length > 0);
-
-        final StringValue actual = Serializer.deserialize(bytes, expected.getDescriptorForType());
-        assertEquals(expected, actual);
-    }
-}
